# documentation/swagger.yaml
...
swagger: '2.0'
info:
  title: 1ClickTrips API
  description: Business travel at its best
  version: "1.0.0"
# the domain of the service
host: api.1clicktrips.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /account/profile:
    get:
      summary: User Profile
      description: |
        The User Profile endpoint returns information about the 1clicktrips
        user that has authenticated with the application.
      tags:
        - User
        - Settings
      responses:
        200:
          description: Profile information for a user
          schema:
            $ref: '#/definitions/Profile'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post: 
      summary: Update User Profile
      description: |
        The User Profile can be only updated collectively.
      tags:
        - User
        - Settings
      parameters:
        - name: body
          in: body
          required: true
          schema: 
            $ref: '#/definitions/Profile'
      responses:
        200:
          description: Update of Profile successful
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /account/settings:
    get:
      summary: User Settings
      description: |
        The Settings endpoint returns information on the user settings of 
        the currently authenticated user. 
      tags:
        - Settings
        - User
      responses:
        200:
          description: An array of settings
          schema:
            type: array
            items:
              $ref: '#/definitions/Setting'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post: 
      summary: Update User Settings
      description: |
        User settings can be updated one at a time. It must be assured, that
        all information are supplied to be able to create a new setting or 
        update an existing one.
      tags:
        - Settings
        - User
      parameters:
        - name: body
          in: body
          required: true
          schema: 
            $ref: '#/definitions/Setting'
      responses:
        200:
          description: Update of setting successful
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /account/favorites:
    get:
      summary: Get User Favorites
      description: |
        The User Favorites endpoint returns data about a user's favorites 
        that he saved before.
        - name: limit
          in: query
          type: integer
          format: int32
          description: |
            Number of items to retrieve. Default is 5, maximum is 100.
      tags:
        - User
      responses:
        200:
          description: List of favorites for the given user
          schema:
            $ref: '#/definitions/Favorites'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post: 
      summary: Update Favorites
      description: |
        Save a new favorite or edit an existing one. 
      tags:
        - User
      parameters:
        - name: body
          in: body
          required: true
          schema: 
            $ref: '#/definitions/Favorite'
      responses:
        200:
          description: Creation/Update of favorites successful
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /account/bookings:
    get:
      summary: Booking History
      description: |
        The Bookings endpoint returns data about a user's lifetime 
        activity with 1clicktrips.
      parameters:
        - name: offset
          in: query
          type: integer
          format: int64
          description: |
            The offset is given as seconds since the UNIX epoche. Only 
            look for trips that have been booked after this date. Default
            is zero.
        - name: limit
          in: query
          type: integer
          format: int32
          description: |
            Number of items to retrieve. Default is 5, maximum is 30.
      tags:
        - User
      responses:
        200:
          description: Booking history for the given user
          schema:
            type: array
            items:
              $ref: '#/definitions/Booking'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  #########################################################################
  # Search API
  #########################################################################
...
definitions:
...
  Setting:
    properties:
      category: 
        type: string
        description: The category of the key-value pair
      key: 
        type: string
        description: The settings key (e.g. early bird)
      value:
        type: string
        description: The settings value (e.g. true)
      description: 
        type: string
        description: The description for the key-value pair
...
  Profile:
    required:
      - email
    properties:
      first-name:
        type: string
        description: First name of the user
      last-name:
        type: string
        description: Last name of the user
      email:
        type: string
        description: Email address of the user
      address:
        $ref: '#/definitions/Address'
      image:
        type: string
        description: Image URL of the user
      promo-code:
        type: string
        description: Promo code of the user
  Address:
    properties:
      street: 
        type: string
      zip:
        type: integer
        format: int32
      city: 
        type: string
      country:
        type: string
  Favorites:
    properties:
      count:
        type: integer
        format: int32
        description: Total number of favorites available
      trips:
        type: array
        items:
          $ref: '#/definitions/Favorite'
  Favorite:
    required: 
      - start
      - end
    properties:
      uuid:
        type: string
        description: Unique identifier for the favorite
      start:
        type: string
        description: |  
          The start location of the booking. Can be an address
          or a description (e.g. home)
      end:
        type: string
        description: |  
          The end location of the booking. Can be an address
          or a description (e.g. company name)
      transport:
        type: array
        items:
          type: number
          format: int32
          description: |
            Travel option used. 
            Walk = 1, Taxi = 2, Car = 4, Bus = 8, 
            Train = 16, Plane = 32
  Booking:
    required:
      - id
      - title
      - destination
      - startDate
      - endDate
      - participants
    properties:
      id:
        type: string
        description: Unique identifier for the booking
      title:
        type: string
        description: A speaking title for the booking
      destination:
        type: string
        description: |  
          The destination location of the booking. Can be an address
          or a description (e.g. company name)
      tags:
        type: array
        items:
          type: string
          description: Tags that can be used to identify the trip.
      startDate:
        type: string
        format: '2015-02-09T02:54:51+00:0'
      endDate:
        type: string
        format: '2015-02-09T02:54:51+00:0'
      reference:
        type: string
        description: | 
          A url to any sort of information for that trip. Could
          be a link to the event or the host company.
      hotel:
        type: object
        $ref: '#/definitions/Hotel'
      participants:
        type: array
        items: 
          type: object
          $ref: '#/definitions/Participant'
  Hotel:
    required:
      - name
    properties:
      name:
        type: string
      checkIn:
        type: string
        format: ""
      url:
        type: string
        decription: Link to the hotel website
      image:
        type: string
        decription: Link to the hotel image
  Participant:
    required:
      - email
      - booked
      - image
    properties:
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      booked:
        type: boolean
      arrivalTime:
        type: string
        description: '2015-02-09T02:54:51+00:0'
      image:
        type: string
        description: Image url for the participant
      profile:
        type: string
        description: Link to a user profile.
...
  Error:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string